Reworked AnimationComponent to work with States
