+++ CSC8508CoreClasses/StateComponent.h
@@ -1,38 +1,29 @@
 #pragma once
-//#include "GameObject.h"
-//#include "PhysicsComponent.h"
-
-#include <map>
-
-#include "IStateTransition.h"
-#include "IState.h"
-#include "IComponent.h"
+#include "GameObject.h"
+#include "PhysicsComponent.h"
 
 
 namespace NCL {
     namespace CSC8508 {
-        //class IState;
-        //class IStateTransition;
-        typedef std::multimap<IState*, IStateTransition*> ITransitionContainer;
-        typedef ITransitionContainer::iterator ITransitionIterator;
-
-        class StateComponent : public IComponent {
+        class StateMachine;
+        class StateComponent : public IComponent  {
         public:
             StateComponent(GameObject& gameObject);
             ~StateComponent();
 
-            void AddState(IState* s);
-            void AddTransition(IStateTransition* t);
+            virtual void Update(float dt);
 
-            virtual void Update(float dt, GameObject& gameObject);
-
+            void OnAwake() override
+            {
+                physics = GetGameObject().TryGetComponent<PhysicsComponent>();
+            }
 
         protected:
-            IState* activeState;
-            
-            std::vector<IState*> allStates;
-            ITransitionContainer allTransitions;
+            void MoveLeft(float dt);
+            void MoveRight(float dt);
 
+            StateMachine* stateMachine;
+            PhysicsComponent* physics;
             float counter;
         };
     }
